# backend/api/src/public/docs/openapi.yaml
openapi: 3.0.0
info:
  title: 'Social Network'
  description: '## An attempt to implement a social network with basic functionality using "Laravel 10" for backend and "React Framework" for frontend.'
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
    description: 'Local API server'
paths:
...
  /api/auth/register:
    post:
      tags:
        - auth
      summary: 'Register a user'
      operationId: api.auth.register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 'Registration was successful'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessRegisterResource'
        '400':
          description: "The passwords don't match"
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: failure }
                  message: { type: string, example: "The passwords don't match" }
                type: object
        '401':
          description: 'Something went wrong while creating token for user'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: failure }
                  message: { type: string, example: 'Something went wrong while creating token for user' }
                type: object
        '422':
          description: 'Validation errors occurred'
          content:
            application/json:
              schema:
                properties:
                  errors: { properties: { first_name: { type: array, items: { type: string, example: 'The first name field is required.' } }, last_name: { type: array, items: { type: string, example: 'The last name field is required.' } }, email: { type: array, items: { type: string, example: 'The email has already been taken.' } }, username: { type: array, items: { type: string, example: 'The username has already been taken.' } }, password: { type: array, items: { type: string, example: 'The password field is required.' } }, password_confirmation: { type: array, items: { type: string, example: 'The password confirmation field is required.' } } }, type: object }
                type: object
        '500':
          description: 'Some internal server error occurred'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResource'
...
components:
  schemas:
...
    RegisterRequest:
      required:
        - first_name
        - last_name
        - email
        - password
        - password_confirmation
      properties:
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john@example.com
        username:
          type: string
          example: john_doe123
        password:
          type: string
          example: '123456'
        password_confirmation:
          type: string
          example: '123456'
      type: object
...
    SuccessRegisterResource:
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: 'Registration was successful'
        user:
          $ref: '#/components/schemas/User'
        authorization:
          $ref: '#/components/schemas/TokenResource'
      type: object
    TokenResource:
      properties:
        access_token:
          type: string
          example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0L2FwaS9hdXRoL3JlZ2lzdGVyIiwiaWF0IjoxNzA2NTUyMTM4LCJleHAiOjE3MDY5MTIxMzgsIm5iZiI6MTcwNjU1MjEzOCwianRpIjoiRk9hclkyZ2VvckhPT1llbSIsInN1YiI6IjQiLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3IiwiZW1haWwiOiJ2YXN5YTJAdmFzeWEuY29tIiwibmFtZSI6InZhc3lhMiJ9.APeVTaT7gnkVVboKkfXDndq1vjQFiiaMPCw21aS9hL8
        type:
          type: string
          example: bearer
      type: object
    InternalServerErrorResource:
      properties:
        status:
          type: string
          example: failure
        message:
          type: string
          example: 'Internal server error'
      type: object
    User:
      properties:
        id:
          type: integer
          example: '2'
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john@example.com
        username:
          type: string
          example: john_doe123
      type: object
  securitySchemes:
    BearerToken:
      type: http
      bearerFormat: JWT
      scheme: bearer